Add HomePageRouter + RouterProvider.
