+++ src/app/router.tsx
@@ -1,0 +1,9 @@
+import { HomePageRoute } from 'routes';
+import { createBrowserRouter } from 'react-router-dom';
+
+export const router = createBrowserRouter([
+  {
+    path: '/',
+    element: <HomePageRoute />,
+  },
+]);
